+++ functions/__init__.py
@@ -1,14 +1,13 @@
 from .messagecolors import MessageColors
-from . import exceptions
-from .embed import embed
-from . import checks
+from . import exceptions, checks, config
+from .myembed import embed
 
+from .custom_contexts import MyContext, MySlashContext
 from .global_cog import GlobalCog
 from .reply import msg_reply
-from .slash_context import MySlashContext
 from .time import timeit
-from .relay_info import relay_info
-from .get_reddit_post import get_reddit_post
+from .relay import relay_info
+from .reddit_post import get_reddit_post
 from .mysql_connection import mydb_connect, query
 dev_guilds = [243159711237537802, 707441352367013899, 215346091321720832]
